Move non-task functions into helpers module
